add info

+++ tensorflow/core/kernels/matmul_op_impl.h
@@ -375,8 +375,8 @@ struct LaunchBatchMatMul<GPUDevice, Scalar> {
     const int64_t batch_size = bcast.output_batch_size();
     auto blas_transpose_a = trans[adj_x ? 2 : (trans_x ? 1 : 0)];
     auto blas_transpose_b = trans[adj_y ? 2 : (trans_y ? 1 : 0)];
-    LOG(INFO) << "blas_transpose_a: " << blas_transpose_a;
-    LOG(INFO) << "blas_transpose_b: " << blas_transpose_b;
+    LOG(INFO) << "blas_transpose_a: " << (int)blas_transpose_a;
+    LOG(INFO) << "blas_transpose_b: " << (int)blas_transpose_b;
 
     auto* stream = context->op_device_context()->stream();
     OP_REQUIRES(context, stream, errors::Internal("No GPU stream available."));
